Attempt to reduce the size of the pyobject.rs files by splitting out builtin types.
