+++ vm/src/exceptions.rs
@@ -3,7 +3,8 @@ use crate::obj::objsequence;
 use crate::obj::objtuple::{PyTuple, PyTupleRef};
 use crate::obj::objtype;
 use crate::obj::objtype::PyClassRef;
-use crate::pyobject::{create_type, IdProtocol, PyContext, PyObjectRef, PyResult, TypeProtocol};
+use crate::pyobject::{IdProtocol, PyContext, PyObjectRef, PyResult, TypeProtocol};
+use crate::types::create_type;
 use crate::vm::VirtualMachine;
 use itertools::Itertools;
 use std::fs::File;
